+++ lib/braintree/customer_gateway.rb
@@ -75,7 +75,9 @@ module Braintree
         ],
       ]
       [
-        :company, :email, :fax, :first_name, :id, :last_name, :phone, :website,
+        :company, :email, :fax, :first_name, :id,
+        {:international_phone => [:country_code, :national_number]},
+        :last_name, :phone, :website,
         :device_data, :payment_method_nonce,
         {:risk_data => [:customer_browser, :customer_ip]},
         {:credit_card => credit_card_signature},
@@ -139,7 +141,9 @@ module Braintree
         ],
       ]
       [
-        :company, :email, :fax, :first_name, :id, :last_name, :phone, :website,
+        :company, :email, :fax, :first_name, :id,
+        {:international_phone => [:country_code, :national_number]},
+        :last_name, :phone, :website,
         :device_data, :payment_method_nonce, :default_payment_method_token,
         {:credit_card => credit_card_signature},
         {:tax_identifiers => [:country_code, :identifier]},