+++ spec/unit/braintree/transaction/customer_details_spec.rb
@@ -11,9 +11,10 @@ describe Braintree::Transaction::CustomerDetails do
         :company => "Smith Co.",
         :website => "http://www.example.com",
         :phone => "6145551234",
+        :international_phone => {:country_code=>"1", :national_number=>"3121234567"},
         :fax => "3125551234",
       )
-      expect(details.inspect).to eq(%(#<id: "id", first_name: "Amy", last_name: "Smith", email: "amy.smith@example.com", company: "Smith Co.", website: "http://www.example.com", phone: "6145551234", fax: "3125551234">))
+      expect(details.inspect).to eq(%(#<id: "id", first_name: "Amy", last_name: "Smith", email: "amy.smith@example.com", company: "Smith Co.", website: "http://www.example.com", phone: "6145551234", international_phone: {:country_code=>\"1\", :national_number=>\"3121234567\"}, fax: "3125551234">))
     end
   end
 end