+++ spec/unit/braintree/transaction_gateway_spec.rb
@@ -40,7 +40,7 @@ describe Braintree::TransactionGateway do
         :shared_shipping_address_id, :shipping_address_id, :shipping_amount,
         :ships_from_postal_code, :tax_amount, :tax_exempt, :three_d_secure_authentication_id,:three_d_secure_token, #Deprecated
         :transaction_source, :type, :venmo_sdk_payment_method_code, #Deprecated
-        :sca_exemption, :currency_iso_code, :exchange_rate_quote_id,
+        :sca_exemption, :currency_iso_code, :exchange_rate_quote_id, :foreign_retailer,
         {:line_items => [:commodity_code, :description, :discount_amount, :image_url, :kind, :name, :product_code, :quantity, :tax_amount, :total_amount, :unit_amount, :unit_of_measure, :unit_tax_amount, :upc_code, :upc_type, :url]},
         {:risk_data => [:customer_browser, :customer_device_id, :customer_ip, :customer_location_zip, :customer_tenure]},
         {:credit_card => [:token, :cardholder_name, :cvv, :expiration_date, :expiration_month, :expiration_year, :number, {:payment_reader_card_details => [:encrypted_card_data, :key_serial_number]}, {:network_tokenization_attributes => [:cryptogram, :ecommerce_indicator, :token_requestor_id]}]},
@@ -139,6 +139,10 @@ describe Braintree::TransactionGateway do
         :ships_from_postal_code,
         :line_items => [:commodity_code, :description, :discount_amount, :image_url, :kind, :name, :product_code, :quantity, :tax_amount, :total_amount, :unit_amount, :unit_of_measure, :unit_tax_amount, :upc_code, :upc_type, :url],
       ])
+    end
+
+    it "creates transaction gateway submit for partial settlement signature" do
+       expect(Braintree::TransactionGateway._submit_for_partial_settlement_signature).to include(:final_capture)
     end
   end
 end