+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 4.21.0
+
+- Add `foreign_retailer` to Transaction
+- Add `international_phone` to `Address` and `Customer`
+- Add `funding_source_description` to PayPalAccount
+- Add missing `GooglePayCard` error code
+- Add `REFUND_FAILED` to Webhook Notification
+- Add `final_capture` to Transaction `_submit_for_partial_settlement_signature`
+- Deprecate `paypal_tracking_id` in favor of `paypal_tracker_id` in `package_details.rb`
+
 ## 4.20.0
 
 - Deprecate `credit` method in `credit_card` and `credit_card_gateway`